update-database

+++ AdoptPet_Server/AdoptPet.Domain/Entities/Donor.cs
@@ -11,6 +11,6 @@ namespace AdoptPet.Domain.Entities
         // khoa ngoai
         public int LocationId { get; set; }
         public string UserId { get; set; } = string.Empty;
-        public int Status { get; set; }
+        public byte Status { get; set; }
     }
 }
